third_party/blink/web_tests/platform/generic/media/media-fragments/TC0004-expected.txt


Title: TC0004
Fragment: 't=0,9.97'
Comment: The media is requested from 0 to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
